+++ src/leftPanel/control/controlWidget.ts
