+++ src/leftPanel/plugin.ts
